@@ -27,6 +27,39 @@ class PhpDumperTest extends TestCase
 {
     public function testDumpContainerWithProxyService()
     {
+        $this->assertStringMatchesFormatFile(
+            __DIR__.'/../Fixtures/php/lazy_service_structure.txt',
+            $this->dumpLazyServiceProjectServiceContainer(),
+            '->dump() does generate proxy lazy loading logic.'
+        );
+    }
+
+    /**
+     * Verifies that the generated container retrieves the same proxy instance on multiple subsequent requests.
+     */
+    public function testDumpContainerWithProxyServiceWillShareProxies()
+    {
+        if (!class_exists('LazyServiceProjectServiceContainer', false)) {
+            eval('?>'.$this->dumpLazyServiceProjectServiceContainer());
+        }
+
+        $container = new \LazyServiceProjectServiceContainer();
+
+        $proxy = $container->get('foo');
+        $this->assertInstanceOf('stdClass', $proxy);
+        $this->assertInstanceOf('ProxyManager\Proxy\LazyLoadingInterface', $proxy);
+        $this->assertSame($proxy, $container->get('foo'));
+
+        $this->assertFalse($proxy->isProxyInitialized());
+
+        $proxy->initializeProxy();
+
+        $this->assertTrue($proxy->isProxyInitialized());
+        $this->assertSame($proxy, $container->get('foo'));
+    }
+
+    private function dumpLazyServiceProjectServiceContainer()
+    {
         $container = new ContainerBuilder();
 
         $container->register('foo', 'stdClass');
@@ -37,38 +70,6 @@ class PhpDumperTest extends TestCase
 
         $dumper->setProxyDumper(new ProxyDumper());
 
-        $dumpedString = $dumper->dump();
-
-        $this->assertStringMatchesFormatFile(
-            __DIR__.'/../Fixtures/php/lazy_service_structure.txt',
-            $dumpedString,
-            '->dump() does generate proxy lazy loading logic.'
-        );
-    }
-
-    /**
-     * Verifies that the generated container retrieves the same proxy instance on multiple subsequent requests.
-     */
-    public function testDumpContainerWithProxyServiceWillShareProxies()
-    {
-        if (class_exists(StaticProxyConstructor::class)) { // detecting ProxyManager v2
-            require_once __DIR__.'/../Fixtures/php/lazy_service_with_hints.php';
-        } else {
-            require_once __DIR__.'/../Fixtures/php/lazy_service.php';
-        }
-
-        $container = new \LazyServiceProjectServiceContainer();
-
-        /* @var $proxy \stdClass_c1d194250ee2e2b7d2eab8b8212368a8 */
-        $proxy = $container->get('foo');
-        $this->assertInstanceOf('stdClass_c1d194250ee2e2b7d2eab8b8212368a8', $proxy);
-        $this->assertSame($proxy, $container->get('foo'));
-
-        $this->assertFalse($proxy->isProxyInitialized());
-
-        $proxy->initializeProxy();
-
-        $this->assertTrue($proxy->isProxyInitialized());
-        $this->assertSame($proxy, $container->get('foo'));
+        return $dumper->dump(array('class' => 'LazyServiceProjectServiceContainer'));
     }
 }
