@@ -113,7 +113,7 @@ class TranslationUpdateCommandTest extends TestCase
                 })
             );
 
-        $loader = $this->getMockBuilder('Symfony\Bundle\FrameworkBundle\Translation\TranslationLoader')->getMock();
+        $loader = $this->getMockBuilder('Symfony\Component\Translation\Loader\TranslationLoader')->getMock();
         $loader
             ->expects($this->any())
             ->method('loadMessages')
