@@ -64,12 +64,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'Symfony\Component\DependencyInjection\Tests\Fixtures\Container33\Foo' service.
+     * Gets the public 'Symfony\Component\DependencyInjection\Tests\Fixtures\Container33\Foo' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\Container33\Foo A Symfony\Component\DependencyInjection\Tests\Fixtures\Container33\Foo instance
+     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\Container33\Foo
      */
     protected function getSymfony_Component_DependencyInjection_Tests_Fixtures_Container33_FooService()
     {
