@@ -72,12 +72,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'bar_service' service.
+     * Gets the public 'bar_service' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getBarServiceService()
     {
@@ -85,12 +82,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'foo_service' service.
+     * Gets the public 'foo_service' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \Symfony\Component\DependencyInjection\ServiceLocator A Symfony\Component\DependencyInjection\ServiceLocator instance
+     * @return \Symfony\Component\DependencyInjection\ServiceLocator
      */
     protected function getFooServiceService()
     {
@@ -104,12 +98,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'translator.loader_1' service.
+     * Gets the public 'translator.loader_1' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getTranslator_Loader1Service()
     {
@@ -117,12 +108,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'translator.loader_2' service.
+     * Gets the public 'translator.loader_2' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getTranslator_Loader2Service()
     {
@@ -130,12 +118,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'translator.loader_3' service.
+     * Gets the public 'translator.loader_3' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getTranslator_Loader3Service()
     {
@@ -143,12 +128,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'translator_1' service.
+     * Gets the public 'translator_1' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator A Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator instance
+     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator
      */
     protected function getTranslator1Service()
     {
@@ -158,12 +140,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'translator_2' service.
+     * Gets the public 'translator_2' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator A Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator instance
+     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator
      */
     protected function getTranslator2Service()
     {
@@ -177,12 +156,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'translator_3' service.
+     * Gets the public 'translator_3' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator A Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator instance
+     * @return \Symfony\Component\DependencyInjection\Tests\Fixtures\StubbedTranslator
      */
     protected function getTranslator3Service()
     {
@@ -199,16 +175,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'baz_service' service.
+     * Gets the private 'baz_service' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * This service is private.
-     * If you want to be able to request this service from the container directly,
-     * make it public, otherwise you might end up with broken code.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getBazServiceService()
     {
