@@ -66,12 +66,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'bar_service' service.
+     * Gets the public 'bar_service' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getBarServiceService()
     {
@@ -79,12 +76,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'foo_service' service.
+     * Gets the public 'foo_service' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getFooServiceService()
     {
@@ -92,16 +86,9 @@ class ProjectServiceContainer extends Container
     }
 
     /**
-     * Gets the 'baz_service' service.
+     * Gets the private 'baz_service' shared service.
      *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * This service is private.
-     * If you want to be able to request this service from the container directly,
-     * make it public, otherwise you might end up with broken code.
-     *
-     * @return \stdClass A stdClass instance
+     * @return \stdClass
      */
     protected function getBazServiceService()
     {
