@@ -60,7 +60,7 @@ EODUMP;
 
         $xDump = <<<'EODUMP'
 Symfony\Component\VarDumper\Caster\ConstStub {
-  +type: "ref"
+  +type: 1
   +class: "2017-08-30 00:00:00.000000 Europe/Zurich (+02:00)"
   +value: """
     Wednesday, August 30, 2017\n
@@ -145,7 +145,7 @@ EODUMP;
 
         $xDump = <<<EODUMP
 Symfony\Component\VarDumper\Caster\ConstStub {
-  +type: "ref"
+  +type: 1
   +class: "$xInterval"
   +value: "$xSeconds"
   +cut: 0
@@ -249,7 +249,7 @@ EODUMP;
 
         $xDump = <<<EODUMP
 Symfony\Component\VarDumper\Caster\ConstStub {
-  +type: "ref"
+  +type: 1
   +class: "$xTimezone"
   +value: "$xRegion"
   +cut: 0
